I have something that looks like a graph!
The code from Mike Bostock was implemented and altered to fit my needs.
I basically wrote evertyhing from scratch having Mike's implementation
as a reference. As for now, the visualization can show a graph of some
nodes that are crated using the script and the id of each node i
written.

Next task will be to make it so that you can add an adjacancy matrix and
it will create the graph using that.